added test to ensure a non owner can not delete a thread

+++ app/Livewire/Threads/ManageThreads.php
@@ -46,6 +46,9 @@ class ManageThreads extends Component
     public function deleteThread($thread)
     {
         $thread = Thread::find($thread);
+        if($thread->user_id != auth()->id()){
+            return redirect('/login');
+        }
         if ($thread->replies()) {
             $thread->replies()->delete();
         }
